Removed stupid stuff
